Triage TODOs and placeholder logic in core code paths
Remove resolved/stale TODO comments, delete dead-code RKink OLS
placeholders in reporting.py, extract shared test utilities to
conftest.py, and create sub-issues (#725, #726, #727) for TODOs
that require future work.

Fixes #666
