+++ causalpy/experiments/diff_in_diff.py
@@ -160,10 +160,9 @@ class DifferenceInDifferences(BaseExperiment):
             }
             self.model.fit(X=self.X, y=self.y, coords=COORDS)
         elif isinstance(self.model, RegressorMixin):
-            # For scikit-learn models, automatically set fit_intercept=False
-            # This ensures the intercept is included in the coefficients array rather than being a separate intercept_ attribute
-            # without this, the intercept is not included in the coefficients array hence would be displayed as 0 in the model summary
-            # TODO: later, this should be handled in ScikitLearnAdaptor itself
+            # Ensure the intercept is part of the coefficients array rather than
+            # a separate intercept_ attribute.  See #726 for centralising this
+            # in ScikitLearnAdaptor.
             if hasattr(self.model, "fit_intercept"):
                 self.model.fit_intercept = False
             self.model.fit(X=self.X, y=self.y)
