+++ causalpy/tests/conftest.py
@@ -19,7 +19,10 @@ Functions:
 """
 
 import numpy as np
+import pandas as pd
 import pytest
+
+import causalpy as cp
 
 # Try to use PyMC's testing helpers if available; otherwise, fall back to no-op fixtures
 try:  # pragma: no cover - conditional import for compatibility across PyMC versions
@@ -59,3 +62,56 @@ def mock_sample_for_doctest(request):
     import pymc as pm
 
     pm.sample = mock_sample
+
+
+def reg_kink_function(x, beta, kink):
+    """Piecewise linear function with a kink, for regression kink design tests."""
+    return (
+        beta[0]
+        + beta[1] * x
+        + beta[2] * x**2
+        + beta[3] * (x - kink) * (x >= kink)
+        + beta[4] * (x - kink) ** 2 * (x >= kink)
+    )
+
+
+def setup_regression_kink_data(kink):
+    """Generate synthetic data for regression kink design tests."""
+    rng = np.random.default_rng(42)
+    N = 50
+    beta = [0, -1, 0, 2, 0]
+    sigma = 0.05
+    x = rng.uniform(-1, 1, N)
+    y = reg_kink_function(x, beta, kink) + rng.normal(0, sigma, N)
+    return pd.DataFrame({"x": x, "y": y, "treated": x >= kink})
+
+
+@pytest.fixture()
+def banks_data():
+    """Load and reshape the banks dataset for DiD integration tests.
+
+    Returns (df_long, treatment_time) where treatment_time has been
+    shifted to 0 so that tests don't depend on absolute years.
+    """
+    treatment_time = 1930.5
+    df = (
+        cp.load_data("banks")
+        .filter(items=["bib6", "bib8", "year"])
+        .rename(columns={"bib6": "Sixth District", "bib8": "Eighth District"})
+        .groupby("year")
+        .median()
+    )
+    df.index = df.index - treatment_time
+    treatment_time = 0
+    df.reset_index(level=0, inplace=True)
+    df_long = pd.melt(
+        df,
+        id_vars=["year"],
+        value_vars=["Sixth District", "Eighth District"],
+        var_name="district",
+        value_name="bib",
+    ).sort_values("year")
+    df_long["unit"] = df_long["district"]
+    df_long["post_treatment"] = df_long.year >= treatment_time
+    df_long = df_long.replace({"district": {"Sixth District": 1, "Eighth District": 0}})
+    return df_long, treatment_time
